Ajout de la fonctionnalité de notification dans le back office depuis /mailbox pour consulter et /compose pour envoyer.

+++ src/main/java/com/saleka/application/notification/client/ClientService.java
@@ -2,8 +2,10 @@ package com.saleka.application.notification.client;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
+
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @Service
@@ -14,7 +16,7 @@ public class ClientService {
     public List<Client> listClients(){
         return clientRepository.findAll();
     }
-    public Client save(Client client){
+    public Client save(@Valid Client client){
         return clientRepository.save(client);
     }
     public Client findById(Long id){
@@ -22,27 +24,29 @@ public class ClientService {
     }
 
     public Client findByEmail(String email){
-        return clientRepository.findByEmail(email);
+        return clientRepository.findByEmail(email).orElseThrow(()->new IllegalStateException("No Client Found"));
     }
-    public Client newClient(Client client){
-        if(client == null){
-            throw new IllegalStateException("Client Not Found");
+    public Client addClient(@NotNull @Valid Client client){
+        if(clientRepository.existsByEmail(client.getEmail())){
+            client.setId(findByEmail(client.getEmail()).getId());
         }
-        clientRepository.saveAndFlush(client);
-        return client;
+        return clientRepository.saveAndFlush(client);
     }
-    public void delete(Long id){
-        clientRepository.deleteById(id);
-    }
-    @Transactional
-    public Client updateClient(Long id, Client client){
-        if(client == null){
-            throw new IllegalStateException("The Update couldn't be resolved the Post object given was null");
-        }
+    public Client addClient(@NotNull @Valid Client client,Long id){
         if(clientRepository.existsById(id)){
             client.setId(id);
-            clientRepository.saveAndFlush(client);
         }
-        return client;
+        return clientRepository.saveAndFlush(client);
+    }
+    public boolean removeClient(@NotNull @Valid Client client){
+        if(clientRepository.deleteClientByEmail(client.getEmail()) == 1){
+            return true;
+        }
+        return false;
+    }
+    public void removeClient(Long id){
+        if(clientRepository.existsById(id)){
+            clientRepository.deleteById(id);
+        }
     }
 }
